+++ swarm/CHANGELOG.md
@@ -22,6 +22,7 @@
 ## 0.43.0
 
 - Allow `NetworkBehaviours` to create and remove listeners.
+- 
   See [PR 3292].
 
 - Raise MSRV to 1.65.